Move ConfigCx into mod kas-core::event::cx
