+++ crates/kas-core/src/event/cx/mod.rs
@@ -9,6 +9,7 @@ use linear_map::{LinearMap, set::LinearSet};
 use smallvec::SmallVec;
 use std::any::TypeId;
 use std::collections::{HashMap, VecDeque};
+use std::fmt::Debug;
 use std::future::Future;
 use std::ops::{Deref, DerefMut};
 use std::pin::Pin;
@@ -359,6 +360,128 @@ impl EventState {
         } else {
             Some(id)
         };
+    }
+}
+
+/// Widget configuration and update context
+///
+/// This type supports access to [`EventState`] via [`Deref`] / [`DerefMut`]
+/// and to [`SizeCx`] via [`Self::size_cx`].
+#[must_use]
+pub struct ConfigCx<'a> {
+    theme: &'a dyn ThemeSize,
+    state: &'a mut EventState,
+    resize: Option<ActionResize>,
+    redraw: Option<ActionRedraw>,
+}
+
+impl<'a> Deref for ConfigCx<'a> {
+    type Target = EventState;
+    fn deref(&self) -> &EventState {
+        self.state
+    }
+}
+impl<'a> DerefMut for ConfigCx<'a> {
+    fn deref_mut(&mut self) -> &mut EventState {
+        self.state
+    }
+}
+
+impl<'a> ConfigCx<'a> {
+    /// Construct
+    #[cfg_attr(not(feature = "internal_doc"), doc(hidden))]
+    #[cfg_attr(docsrs, doc(cfg(internal_doc)))]
+    pub fn new(sh: &'a dyn ThemeSize, ev: &'a mut EventState) -> Self {
+        ConfigCx {
+            theme: sh,
+            state: ev,
+            resize: None,
+            redraw: None,
+        }
+    }
+
+    /// Access a [`SizeCx`]
+    #[inline]
+    pub fn size_cx<'b>(&'b mut self) -> SizeCx<'b>
+    where
+        'a: 'b,
+    {
+        SizeCx::new(self.state, self.theme)
+    }
+
+    /// Configure a widget
+    ///
+    /// All widgets must be configured after construction; see
+    /// [widget lifecycle](crate::Widget#widget-lifecycle) and
+    /// [configuration](Events#configuration).
+    /// Widgets must always be sized after configuration.
+    ///
+    /// Assigns `id` to the widget. This must be valid and is usually
+    /// constructed with [`Events::make_child_id`].
+    #[inline]
+    pub fn configure(&mut self, mut widget: Node<'_>, id: Id) {
+        // This recurses; avoid passing existing state in
+        // (Except redraw: this doesn't matter.)
+        let start_resize = std::mem::take(&mut self.resize);
+        widget._configure(self, id);
+        self.resize = self.resize.or(start_resize);
+    }
+
+    /// Update a widget
+    ///
+    /// All widgets must be updated after input data changes; see
+    /// [update](Events#update).
+    #[inline]
+    pub fn update(&mut self, mut widget: Node<'_>) {
+        // This recurses; avoid passing existing state in
+        // (Except redraw: this doesn't matter.)
+        let start_resize = std::mem::take(&mut self.resize);
+        widget._update(self);
+        self.resize = self.resize.or(start_resize);
+    }
+
+    /// Set a target for messages of a specific type when sent to `Id::default()`
+    ///
+    /// Messages of this type sent to `Id::default()` from any window will be
+    /// sent to `id`.
+    pub fn set_send_target_for<M: Debug + 'static>(&mut self, id: Id) {
+        let type_id = TypeId::of::<M>();
+        self.pending_send_targets.push((type_id, id));
+    }
+
+    /// Notify that a widget must be redrawn
+    ///
+    /// "The current widget" is inferred from the widget tree traversal through
+    /// which the `EventCx` is made accessible. The resize is handled locally
+    /// during the traversal unwind if possible.
+    #[inline]
+    pub fn redraw(&mut self) {
+        self.redraw = Some(ActionRedraw);
+    }
+
+    /// Require that the current widget (and its descendants) be resized
+    ///
+    /// "The current widget" is inferred from the widget tree traversal through
+    /// which the `EventCx` is made accessible. The resize is handled locally
+    /// during the traversal unwind if possible.
+    #[inline]
+    pub fn resize(&mut self) {
+        self.resize = Some(ActionResize);
+    }
+
+    #[inline]
+    pub(crate) fn needs_redraw(&self) -> bool {
+        self.redraw.is_some()
+    }
+
+    #[inline]
+    pub(crate) fn needs_resize(&self) -> bool {
+        self.resize.is_some()
+    }
+
+    #[inline]
+    pub(crate) fn set_resize(&mut self, resize: Option<ActionResize>) {
+        self.resize = resize;
     }
 }
